Convert config to a custom resource
Also I’m pretty sure this never worked. Now we’re testing it
