+++ test/integration/source/git_installed_spec.rb
@@ -1,6 +1,4 @@
-describe 'git::default' do
-  describe command('/usr/local/bin/git --version') do
-    its(:exit_status) { should eq 0 }
-    # its(:stdout) { should match(/something/) }
-  end
+describe command('/usr/local/bin/git --version') do
+  its(:exit_status) { should eq 0 }
+  # its(:stdout) { should match(/something/) }
 end
